% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{calculate_group_transition_freq}
\alias{calculate_group_transition_freq}
\title{calculate_group_transition_freq}
\usage{
calculate_group_transition_freq(x, var)
}
\arguments{
\item{x}{dataframe of JH4 pipeline output}

\item{var}{column name to group by}
}
\description{
calculate SNP transition frequency per group
}
