% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{calculate_mu_freq}
\alias{calculate_mu_freq}
\title{calculate_mu_freq}
\usage{
calculate_mu_freq(x, var)
}
\arguments{
\item{x}{dataframe of JH4 pipeline output}

\item{var}{column name to group by}
}
\description{
calculate SHM frequency
}
