% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{count_snp}
\alias{count_snp}
\title{count_snp}
\usage{
count_snp(x, absolute = T)
}
\arguments{
\item{x}{dataframe of JH4 pipeline output}

\item{absolute}{defaults to TRUE to return SNP counts from all reads, else return SNP counts from each sequence}
}
\description{
count number of SNP
}
