% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{count_transitions}
\alias{count_transitions}
\title{count_transitions}
\usage{
count_transitions(conversions, count)
}
\arguments{
\item{conversions}{conversion column}

\item{count}{count column}
}
\description{
count the number of SNPs for transition frequency
}
