% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcj4m-package.R
\docType{package}
\name{GCJ4M-package}
\alias{GCJ4M}
\alias{GCJ4M-package}
\title{GCJ4M: Germinal Center JH4 Somatic Hypermutation Analysis}
\description{
GCJ4M provides methods to analysis JH4 amplicon sequencing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hungms.github.io/GCJ4M/}
}

}
\author{
\strong{Maintainer}: Miu Shing Hung \email{matthewhungms@gmail.com} (\href{https://orcid.org/0009-0004-0412-4374}{ORCID})

}
\keyword{internal}
