% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{plot_mu_prop}
\alias{plot_mu_prop}
\title{plot_mu_prop}
\usage{
plot_mu_prop(x, var, ...)
}
\arguments{
\item{x}{dataframe of JH4 pipeline output}

\item{var}{column name to group by}
}
\description{
plot proportion of reads with SHM
}
