% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{plot_multi_qc}
\alias{plot_multi_qc}
\title{plot_multi_qc}
\usage{
plot_multi_qc(x, sample = "sample", group = NULL)
}
\arguments{
\item{x}{dataframe of JH4 pipeline output}

\item{sample}{column name indicating sample}

\item{filter}{whether to remove indel sequences in the final output}
}
\description{
plot quality control parameters
}
