% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{plot_seq_copy}
\alias{plot_seq_copy}
\title{plot_seq_copy}
\usage{
plot_seq_copy(x, var, min.copy = 1)
}
\arguments{
\item{x}{dataframe of JH4 pipeline output}

\item{var}{column name to group by}
}
\description{
plot number of reads for each unique sequence
}
