% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{plot_seq_length}
\alias{plot_seq_length}
\title{plot_seq_length}
\usage{
plot_seq_length(x, sample = "sample", group = NULL, ...)
}
\arguments{
\item{x}{dataframe of JH4 pipeline output}

\item{sample}{column name indicating sample}

\item{group}{column name indicating group}
}
\description{
plot the sequencing length from each sample
}
