% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{plot_snp_heatmap}
\alias{plot_snp_heatmap}
\title{plot_snp_heatmap}
\usage{
plot_snp_heatmap(snp_matrix_list)
}
\arguments{
\item{snp_matrix_list}{a list of SNP matrix}
}
\description{
plot heatmap for SNP count matrices
}
