% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{remove_indels}
\alias{remove_indels}
\title{remove_indels}
\usage{
remove_indels(x, sample = "sample", filter = T)
}
\arguments{
\item{x}{dataframe of JH4 pipeline output}

\item{sample}{column name indicating sample}

\item{filter}{whether to remove indel sequences in the final output}
}
\description{
remove sequences with insertions and deletions
}
