% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{remove_lowqc}
\alias{remove_lowqc}
\title{remove_lowqc}
\usage{
remove_lowqc(x, sample = "sample", min.reads = 400, min.copy = 1)
}
\arguments{
\item{x}{dataframe of JH4 pipeline output}

\item{sample}{column name indicating sample}

\item{min.reads}{minimum number of reads; defaults to 400}

\item{min.copy}{minimum reads for each sequence; defaults to 1}
}
\description{
remove samples that do not pass quality control
}
