% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{calculate_group_snp}
\alias{calculate_group_snp}
\title{calculate_group_snp}
\usage{
calculate_group_snp(sample_snp, group_vec = NULL)
}
\arguments{
\item{sample_snp}{a list of count matrix per sample}

\item{group_vec}{a vector of groups for each sample; vector must be same length as the list of snptable}
}
\description{
calculate SNP count matrix per group
}
