% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{calculate_sample_snp}
\alias{calculate_sample_snp}
\title{calculate_sample_snp}
\usage{
calculate_sample_snp(x, sample = "sample", absolute = T)
}
\arguments{
\item{x}{dataframe of JH4 pipeline output}

\item{sample}{column name indicating sample}

\item{absolute}{defaults to TRUE to return SNP counts from all reads, else return SNP counts from unique sequences}
}
\description{
calculate SNP count matrix per sample
}
