% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{count_transitions}
\alias{count_transitions}
\title{count_transitions}
\usage{
count_transitions(conversions, count, absolute = T)
}
\arguments{
\item{conversions}{}

\item{count}{}

\item{absolute}{defaults to TRUE to return SNP counts from all reads, else return SNP counts from unique sequences}
}
\description{
count the number of SNPs for transition frequency
}
