% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{plot_snp_heatmap}
\alias{plot_snp_heatmap}
\title{plot_snp_heatmap}
\usage{
plot_snp_heatmap(snp_list, legend.title = "CPM")
}
\arguments{
\item{snp_list}{a list of SNP matrix}

\item{legend.title}{specify legend title}
}
\description{
plot heatmap for SNP count matrices
}
